+++ re_contact_book_system/frontend/src/main.js
@@ -1,8 +1,14 @@
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
-
 import App from './App.vue'
 import router from './router'
+
+// Bootstrap CSS
+import 'bootstrap/dist/css/bootstrap.min.css'
+import 'bootstrap-icons/font/bootstrap-icons.css'
+
+// Bootstrap JS
+import 'bootstrap/dist/js/bootstrap.bundle.min.js'
 
 const app = createApp(App)